pyproject.toml
[project]
name = "cloud-messaging-cli"
version = "0.1.0"
description = "A command-line tool for managing cloud messaging services (Pub/Sub, SNS/SQS, etc.)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-bigquery>=3.35.0",
    "google-cloud-pubsub>=2.31.0",
    "google-cloud-storage>=3.2.0",
    "tabulate>=0.9.0",
]
